src_scrappers/requirements.txt
scrapy
mysql
sqlalchemy
alembic
python-dotenv